refactor in process dataclasses seperated, putting results into async queue
